Move notebook specific Arrow-encoding
